+++ NEWS.md
@@ -10,6 +10,8 @@
 
 * The vendored [nanoarrow](https://github.com/apache/arrow-nanoarrow) has been updated to its release [0.5.0](https://github.com/apache/arrow-nanoarrow/releases/tag/apache-arrow-nanoarrow-0.5.0) (#733)
 
+* Fragments can now be removed by supplying a vector of fragment URIs (#734)
+
 ## Build and Test Systems
 
 * The nighly valgrind matrix now includes release 2.25.0 (#729)